Changed long integers to those in cstdint.

+++ 003.cpp
@@ -1,10 +1,11 @@
 #include <vector>
 #include <iostream>
+#include <cstdint>
 #include "prime_util.h"
 
 int main()
 {
-    unsigned long long n = 600851475143;
+    uint64_t n = 600851475143;
 
     int prime = 1;
     while (n > 1)
