+++ 007.cpp
@@ -1,10 +1,11 @@
 #include <iostream>
 #include <vector>
+#include <cstdint>
 #include "prime_util.h"
 
 int main()
 {
-    unsigned long long prime = 1;
+    uint64_t prime = 1;
     for (int i = 0; i < 10001; i++)
         prime = util::get_next_prime(prime);
