+++ 010.cpp
@@ -1,12 +1,13 @@
 #include <iostream>
 #include <vector>
+#include <cstdint>
 #include "prime_util.h"
 
 int main()
 {
     std::vector<int> primes = util::get_primes(1999999);
 
-    unsigned long long sum = 0;
+    uint64_t sum = 0;
     for (int &prime : primes)
         sum += prime;
